Adding the ability to disable the email for user testing.
This needed a small hack to re-read the environment setting at request processing time rather than just when the application is configured because that happens just once when the RCMServer is loaded.  After that the yaml is never re-read and all the tests run in a single Ruby virtual machine.  The code now has a hack to re-read that environment variable